requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage
Sphinx
django-extensions
Werkzeug

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar
